Update Segmentation of melody
In music structural analysis, a musical work is decomposed (segmented) into its constituent elements. Segmentation is mostly used in a purely musical-theoretical sense to determine how a piece of music is constructed, how structural units interact, and how they are connected \parencite{Ahlback07}. Due to the ambiguity of music, determining segment boundaries may be challenging \parencite{Margulisetal17}. Even though two persons have similar experience of a musical work, their perceptions of the musical piece or its structural parts may be quite different.

The majority of research on the mechanisms underlying perceptual grouping in music is conducted on adults. Only a few studies have examined how very young toddlers and children of particular age perceive and organize musical structure.\footnote{Throughout this thesis, the term ``children'' is used in accordance with the United Nations Convention on the Rights of the Child - UNCRC 1989 - which defines a ``child'' as any person under the age of 18.} Furthermore, most of research on children is either focused on a certain age group or on specific musical dimensions and features, with little attention paid to music segmentation.

There is currently no research on how children of various ages perceive boundaries in music or how music is segmented. The purpose of this research is to bridge the gap in this population's comprehension of music segmentation and to present a more complete picture of human segmentation.